Use a single terminal interface across multiple hubs to fix bugs with hub restart.
Split out the actual JLine-based terminal interaction to a singleton class
used by subsequent instances of JLineTerminalIO so that hub restart within
a single process doesn't cause terminal errors.
